Remove yarn link:bin script
To adopt TypeScript, we previously used ts-node [1] in [2] as our
shebang. However this required user to install ts-node as a sibling
global dependency and effectively stopped using transpiled JavaScript
files.

So in [3] we reverted back to node to fix the issue in distribution.
However locally when using `yarn/npm link`, it still links that very
same script to global binary directory. Since locally we don't keep the
JavaScript files (we only transpile them when packaging/publishing), it
rendered the `yarn link:bin` setup broken.

As it stands now, it's easier and more straightforward to rely on `yarn
dev` which feeds the `run` script into ts-node, overriding its shebang.

[1] https://github.com/TypeStrong/ts-node
[2] https://github.com/zendesk/zcli/commit/2b799dcc33a8aabdc092010a937619755a60577a
[3] https://github.com/zendesk/zcli/commit/0f59937e13538120c37693637508736c77e8f582