fix: Pin all oclif dependencies to exact versions
As much as we want to allow customers to get latest patches as soon as
possible, oclif on the other hand keeps breaking the Semantic Versioning
2.0.0 [1] contract, where breaking changes must lead to major version
bumps.

After fixing last breakage in [2][3] by pinning minor versions,
oclif/plugin-not-found also upgraded to oclif/core v2 [4] and this time
only patch version was bumped [5].

So being cautious this time and just pin all oclif dependencies to
exactly the same versions as those in the lock file.

[1] https://semver.org
[2] https://github.com/zendesk/zcli/pull/157
[3] https://github.com/zendesk/zcli/commit/3c73b5334e5604ba5c3ed40eb8eb76a758b3c0c3
[4] https://github.com/oclif/plugin-not-found/commit/b3176628b2dddd69c4419f8270fbb510231d3f8e
[5] https://github.com/oclif/plugin-not-found/releases/tag/2.3.17