Bump typescript and `@typescript-eslint`
Let's bump our typescript tooling to something a bit more recent, there are
also a few security vulnerabilities associated to earlier `@typescript-eslint`
which are fixed with this change.

Note: I've used a bit of a hammer to silence linting errors from `yarn lint`,
there may be an argument to revisit linting in a future change to finesse
things a tad.

In short,
- Bump `@typescript-eslint` from 5.30.0 to 7.14.1
- Bump typescript from 4.7.4 to 5.4.5. (I opted for the 5.4.x branch since
  5.5.x seems a bit too recent and not mature yet)
- We also need to bump eslint-config-oclif-typescript from 1.0.2 to 3.1.8, this
  is because it has a transitive dependency on an earlier version of
  `@typescript-eslint` (^4.31.2) [1] that is incompatible with typescript 5.x+ [2]
- Squelch linting errors by switching off no-explicit-any in .eslintrc, and
  also silence unused var errors in test and application code.

[1] https://github.com/oclif/eslint-config-oclif-typescript/blob/v1.0.2/package.json#L8-L9
[2] https://stackoverflow.com/a/77124060