feat: add ability to use ZENDESK_APP_ID env var during apps:update
Someones we want to have seperate apps for staging and production. In these cases they will have a seperate app_id.

In this change we will allow users to set the `ZENDESK_APP_ID` environment variable to specify which app_id to use. This will allow developers to set this in their CICD pipelines.

Fixes: https://github.com/zendesk/zcli/issues/173